Update watch.js to work with windows git bash
In git bash with `const srcDirPrefix = pathLib.resolve("./src")`  it was resolving to `'C:\\projects\\livestream\\src'` when adding the trailing '/' the key it was trying to delete would be `C:\projects\\livestream\\src/` which didn't exist, but if you remove the trailing slash it does exist.  Looking at the code it appears that the trailing slash shouldn't matter unless you have another folder/file who's cache you don't want deleted that starts with `src` or `node_modules` so this shouldn't be run into in practice.